pyproject.toml
[project]
name = "akshar-benchmark"
version = "0.1.0"
description = "Zero-shot Indic Handwriting OCR benchmark with Gemma 4"
requires-python = ">=3.10"
dependencies = [
    "datasets>=2.19.0",
    "Pillow>=10.0.0",
    "tqdm>=4.66.0",
    "huggingface-hub>=0.22.0",
    "transformers>=4.50.0",
    "accelerate>=1.3.0",
    "bitsandbytes>=0.45.0",
    "numpy",
    "sentencepiece",
    "protobuf",
]

[tool.uv]
# PyTorch with CUDA 12.4 — install separately or via override below
# uv pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu124" }]
torchvision = [{ index = "pytorch-cu124" }]
